Factor state mapping logic into seperate function
